Removed System.out.println message.
This message is continuously being logged to the console while Activiti Explorer is running and tasks are being browsed. It looks like a debug message that was left over from development, listing the number of attachments a certain task has. I have removed it and not replaced it by logging since it seems to have no real value as far as logging is concerned and the rest of the class also doesn't appear to do any logging of its own.