Ported https://github.com/Activiti/Activiti/commit/cc14630f18bf8bd5e960308302a1aac1d0a13b4b to v6
(Async executor improvement: before, when executing an exclusive job and when not able to get the lock, the executing thread would be stalled for an amount of time and the locking would be retried until the process instance can be locked. This is now changed to not wait+retry, but 'release' the job again (similar to the 'queue full' behavior), thus making the executing thread free again. This also means that the 'retryWaitTimeInMillis' is now deprecated as it is not used anymore and will be removed in a future version.)