Fixes ACT-4117 : Signal and Boundary event subscription are lost when deploying a new process definition version + ACT-4117 : Signal and Boundary event subscription are lost when deploying a new process definition version
Two new columns were added to the event subscription table + the queries were changed.
Before, when deploying a new process definition, all event subscriptions were deleted (including boundary event subscriptions!) and new event subscriptions for the new version were inserted.
This commit changes this: the event subscriptions will not be deleted anymore, but a flag 'isLatest' will be set to false for previous versions. The current version has isLatest set to true.
This way, when deleting deployments, the previous process definition is looked up, and the matching event subscriptions are set to isLatest=true again.

This commit does not include upgrade scripts. That will follow in a later commit (but first want to verify behaviour on all databases).