ACT-4116 Custom subprocess parser
I tried to get it working with the existing Designer code, but got to
where the next id was determined for each BPMN element within the UI and
it was looping through and finding the next one. Wouldn't it be better
to use a UUID or keep a key list somewhere with the next key value?
Don't you use ids or sids for the web-based design tool already?

Added update for files without existing BPMNDI data.