identitylinks user & group docs did not match implementation
The RestIdentityLink class, which is being used for serialization
in identity link creation and retrieval, has the properties,
"group", and "user".  The doc used "groupId" and "userId" instead.
Creating identity links using the exact doc sample syntax resulted
in the error,
"A group or a user is required to create an identity link."