Trigger process deployed events (#2174)
* Trigger process deployed events

The events will be sent when the Spring Boot application is ready (ApplicationReadyEvent). A ProcessDeployedEvent will be triggered for every process deployed every time the application starts. Listeners to these events should be able do deal with duplicates.

Refs https://github.com/Activiti/Activiti/issues/2149

* Fix transitive dependencies

* Fix integration test

events coming from `WebEnvironment.MOCK` are ignored. So we need to use MOCK here.