AAE-23532 Fix race condition for service tasks using connectors (#4726)
Having `DefaultServiceTaskBehavior` as an `ActivityBehavior` was causing racing conditions because it's instantiated as a singleton. Given the fact that DefaultServiceTaskBehavior is always wrapped inside a `ServiceTaskDelegateExpressionActivityBehavior` that is already an `ActivityBehavior`, there is no reason to have `DefaultServiceTaskBehavior` also implementing `ActivityBehavior`. This PR is simplifying the layers and extracting new interface to be used by `DefaultServiceTaskBehavior`.