Fix wrong selection of MessageStartEvent when process has more than one StartEvent
Issue found when starting a process with multiple MessageStartEvents.
The code for starting a process by a BPMN message also transforms
the MessageEventDefinition's `messageRef` into the associated Message's
name when the `messageRef` doesn't match the passed `messageName`.
This additional logic is wrongly selecting the current StartEvent
with rechecking the new `messageRef` against the `messageName`.

Example scenario:
- Process `x` has to message start events:
  - `startEvt1` with messageEventDefinition with `messageRef="message1Id"`;
  - `startEvt2` with messageEventDefinition with `messageRef="message2Id"`;
- Process `x` also has messages:
  - `<message id="message1Id" name="message1"/>`
  - `<message id="message2Id" name="message2"/>`
- A new process is started with `runtimeService.startProcessInstanceByMessage("message2")`;
- The process starts from `startEvt1` instead of `startEvt2`.

This bug does not happen every time. If only happens when the MessageEventDefinitions'
messageRefs haven't already been transformed by some other part of the codebase.